Changes Multiple File Uploading Feature and docs entity

+++ src/users/users.service.ts
@@ -2,18 +2,24 @@ import { Injectable, HttpException, HttpStatus } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository, DeleteResult, Not } from 'typeorm';
 import { CreateUserDto, UpdateUserDto } from './dto/user.dto';
+import { Docs } from './entity/docs.entity';
 import { User } from './entity/user.entity';
 import { UserRO } from './ro/user.ro';
 
 @Injectable()
 export class UsersService {
-    constructor(@InjectRepository(User) private readonly userRepository: Repository<User>) { }
+    constructor(@InjectRepository(User) private readonly userRepository: Repository<User>, @InjectRepository(Docs) private readonly docsRepository: Repository<Docs>) { }
 
     public async getUsers(): Promise<User[]> {
         return await this.userRepository.find();
     }
 
     public async getUser(_id: string): Promise<UserRO | null> {
+        const user = await this.userRepository.findOne(_id);
+        if (!user) {
+            throw new HttpException("User doesn't exist", HttpStatus.BAD_REQUEST);
+        }
+
         return (await this.userRepository.findOneOrFail(_id)).toResponseObject();
     }
     
@@ -47,6 +53,23 @@ export class UsersService {
     }
 
     public async deleteUser(_id: string): Promise<DeleteResult> {
+        const user = await this.userRepository.findOne(_id);
+        if (!user) {
+            throw new HttpException("User doesn't exist", HttpStatus.BAD_REQUEST);
+        }
+        
         return await this.userRepository.delete(_id);
+    }
+    
+    public async uploadDocs(_id: string, docsResponse) {
+        const user = await this.userRepository.findOne(_id);
+        if (!user) {
+            throw new HttpException("User doesn't exist", HttpStatus.BAD_REQUEST);
+        }
+
+        const docs = this.docsRepository.create();
+        docs.user_id = _id;
+        docs.docsResponse = docsResponse;
+        return await this.docsRepository.save(docs);
     }
 }
